requirements.txt
pytest==7.4.4 
python-dotenv==1.0.1 
requests==2.31.0